enforce JPG only restriction for photo uploads
I did this because JPG images are generally lighter than PNGs, which helps them load faster on load. Also to minimise sanity storage and sanity bandwidth usage (this project is on sanity free plan, which has limitted ressources). Also, enforcing JPG allows to apply Sanity's built-in image compression by appending  to the image URL. This was mentioned in commit , this feature can't be done with PNGs